requirements.txt
torchvision>=0.9.1
pytorch>=1.8.1
sklearn
Pillow>=8.2.0
opencv-python>=4.5.1
matplotlib>=3.1.3
numpy>1.21.2
tqdm